Use "oldest-supported-numpy" instead of plain numpy
This changes the build to use the oldest supported numpy version. With this, the resulting binary will be compatible with all newer NumPy version. This is quite helpful as NumPy's (ABI) compatibility is so that it will support any NumPy version that is newer than the one used during the build.